fix: forgot to rename others
